fix: move JVM-only virtual time tests to jvmTest source set
- Move multipleObserversSameCharacteristic to jvmTest (ObservationReconnectionJvmTest)
- Move collectorCancellationDisablesCccdWhenNoCollectorsRemain to jvmTest (ObservationReconnectionJvmTest)
- Remove @Ignore and JVM-only virtual time code from commonTest

These tests use kotlinx.coroutines.test virtual time APIs (runTest, advanceUntilIdle, testScope.testDispatcher) which are JVM-only and don't work on Kotlin Native (iOS).
